requirements.txt
# Core dependencies
numpy>=1.24.0
scipy>=1.10.0
pyyaml>=6.0

# Machine Learning
torch>=2.0.0
transformers>=4.30.0
tokenizers>=0.13.3
sentencepiece>=0.1.99

# Data Processing
nltk>=3.8
scikit-learn>=1.3.0
pandas>=2.0.0

# Compression
lz4>=4.3.2

# Web Scraping (for corpus building)
requests>=2.31.0
beautifulsoup4>=4.12.0
wikipedia>=1.4.0
urllib3>=2.0.0

# Progress and Utilities
tqdm>=4.65.0
colorama>=0.4.6
click>=8.1.0

# Evaluation
sacrebleu>=2.3.1
rouge-score>=0.1.2

# TensorFlow Lite (for export)
tensorflow>=2.13.0

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0  # Optional

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.7.0
flake8>=6.1.0
mypy>=1.4.0

# Ollama API Client
ollama>=0.1.0  # pip install ollama

# JSON Lines
jsonlines>=3.1.0
